Renaming EightWarps to EightWaves for consistency
