+++ projects/composablekernel/test/ck_tile/gemm_block_scale/test_gemm_quant_fixtures.hpp
@@ -1276,7 +1276,7 @@ class TestCkTileGemmABQuant : public TestCkTileGemmQuantBase<Tuple, TestCkTileGe
 
             using GemmPipeline = std::conditional_t<
                 eight_waves,
-                ck_tile::ABQuantGemmPipelineAgBgCrEightWarps<PipelineProblem>,
+                ck_tile::ABQuantGemmPipelineAgBgCrEightWaves<PipelineProblem>,
                 std::conditional_t<PreshuffleB,
                                    ck_tile::WPABQuantBPipelineAgBgCrV2<PipelineProblem>,
                                    ck_tile::ABQuantGemmPipelineAgBgCrCompV3<PipelineProblem>>>;
